feat: Integrate Popmotion and Parcel for code modularity
This commit adds code changes to integrate two different ways of incorporating modules of code into the project. The first option is to use the Popmotion library for creating animations. The second option is to use Parcel, a tool that bundles all the code together into a single package. Both options provide flexibility and modularity to the project.